Fixed bug JUM-80: Need to refactor some classes related to validations to make them inheritable
