+++ utilities/src/main/java/org/jumbune/yarn/utils/NMTokenCache.java
@@ -1,4 +1,4 @@
-package org.jumbune.utils.yarn;
+package org.jumbune.yarn.utils;
 
 import java.util.concurrent.ConcurrentHashMap;
