+++ utilities/src/main/java/org/jumbune/yarn/utils/RackResolver.java
@@ -1,4 +1,4 @@
-package org.jumbune.utils.yarn;
+package org.jumbune.yarn.utils;
 
 import java.util.ArrayList;
 import java.util.List;
